refactor: aes_ctr -> aesctr and headers & static libs placeholders in esp32 examples
