+++ examples/atclient_esp32_static_components/components/atclient/include/atclient/.gitkeep
@@ -1,0 +1,1 @@
+Copy and paste headers into here
