--- examples/atclient_esp32_static_no_components/main/atclient/include/atclient/at_logger.h
@@ -1,9 +1,0 @@
-#ifndef ATLOGGER_H
-#define ATLOGGER_H
-
-#include <stddef.h>
-
-int atlogger_log(const char *title, const char *message);
-int atlogger_logx(const char *title, const unsigned char *bytes, size_t byteslen);
-
-#endif  // ATLOGGER_H
