+++ packages/atchops/include/atchops/constants.h
@@ -1,7 +1,7 @@
 #ifndef ATCHOPS_CONSTANTS_H
 #define ATCHOPS_CONSTANTS_H
 
-#define ATCHOPS_RNG_PERSONALIZATION "atchops"
+#define ATCHOPS_RNG_PERSONALIZATION "@atchops12345"
 #define ATCHOPS_IV_SIZE 16 // ivs are always 16 bytes long
 
 #endif
