+++ packages/atchops/include/atchops/rsakey.h
@@ -1,25 +1,25 @@
 #ifndef ATCHOPS_RSAKEY_H
 #define ATCHOPS_RSAKEY_H
 
-typedef struct rsakey_param
+typedef struct atchops_rsakey_param
 {
     unsigned long len;    // length of the number in bytes
     unsigned char *value; // hex byte array of the number
-} rsakey_param;
+} atchops_rsakey_param;
 
 typedef struct atchops_rsakey_publickey
 {
-    rsakey_param n; // modulus
-    rsakey_param e; // public exponent
+    atchops_rsakey_param n; // modulus
+    atchops_rsakey_param e; // public exponent
 } atchops_rsakey_publickey;
 
 typedef struct atchops_rsakey_privatekey
 {
-    rsakey_param n; // modulus
-    rsakey_param e; // public exponent
-    rsakey_param d; // private exponent
-    rsakey_param p; // prime 1
-    rsakey_param q; // prime 2
+    atchops_rsakey_param n; // modulus
+    atchops_rsakey_param e; // public exponent
+    atchops_rsakey_param d; // private exponent
+    atchops_rsakey_param p; // prime 1
+    atchops_rsakey_param q; // prime 2
 } atchops_rsakey_privatekey;
 
 void atchops_rsakey_init_publickey(atchops_rsakey_publickey *publickey);
